+++ src/components/Content/ProductCardPopup/ProductCardPopup.less
@@ -1,8 +1,8 @@
-@import "./../../../common/variables";
+@import "./../../common/variables";
 
 .popup {
   width: 280px;
-  border: 1px solid;
+  box-shadow: 0px 25px 35px rgba(0, 0, 0, 0.1);
   position: absolute;
   z-index: 100;
   top: -21px;
