+++ src/components/OrderForm/OrderForm.less
@@ -77,9 +77,6 @@
 
 .form {
   position: fixed;
-  /*   top: 0;
- */
-  z-index: 10000;
   overflow: auto;
   transition: all 300ms linear;
   top: 50%;
