+++ src/components/common/Button/Button.less
@@ -31,7 +31,7 @@
   &_disabled {
     cursor: none;
     pointer-events: none;
-    color: @color-light-grey;
+    opacity: 0.5;
   }
 
   &:hover,
@@ -48,10 +48,6 @@
 
 .button_submit {
   width: 100%;
-
-  &_disabled {
-    opacity: 0.5;
-  }
 }
 
 .button_form {
