+++ src/components/common/Input/Input.less
@@ -6,7 +6,6 @@
 
   .input {
     box-sizing: border-box;
-
     width: 100%;
     height: @control-height;
     display: inline-block;
