+++ src/components/common/Modal/Modal.less
@@ -6,6 +6,7 @@
   transform: translate(-50%, -50%);
   width: 100%;
   height: 100vh;
+  z-index: 10000;
 
   .popup {
     top: 50%;
